requirements.txt
groq
python-dotenv
rich
streamlit
